+++ src/index.ts
@@ -8,7 +8,7 @@ export * from './inputs/Hidden'
 export * from './inputs/ItemSelector'
 export * from './inputs/ItemText'
 export * from './inputs/KeyValue'
-export * from './inputs/MultiSelect2'
+export * from './inputs/MultiSelect'
 export * from './inputs/NumberInput'
 export * from './inputs/Radio'
 export * from './inputs/Select'
