+++ wizards/Placement/Placement.tsx
@@ -5,7 +5,7 @@ import { Fragment, useMemo } from 'react'
 import { ArrayInput, EditMode, Hidden, KeyValue, NumberInput, TextInput } from '../../src'
 import { useEditMode } from '../../src/contexts/EditModeContext'
 import { useItem } from '../../src/contexts/ItemContext'
-import { MultiSelect } from '../../src/inputs/MultiSelect2'
+import { MultiSelect } from '../../src/inputs/MultiSelect'
 import { IResource } from '../common/resource'
 import { IClusterSetBinding } from '../common/resources/IClusterSetBinding'
 import { IPlacement, PlacementKind, PlacementType, Predicate } from '../common/resources/IPlacement'
